Move IDragDropManager to correct namespace
